working on items

+++ src/main/resources/migrations/V1__Initialize_database.sql
@@ -25,7 +25,7 @@ CREATE TABLE items
 (
     id          SERIAL PRIMARY KEY,
     name        varchar,
-    price       numeric,
+    price       integer,
     description varchar,
     item_type   item_type,
     created_at  TIMESTAMP DEFAULT CURRENT_TIMESTAMP
